// client/packages/editor-oss/src/controls/AiWorldController/AiWorldController.types.ts
export enum COMMANDS {
    ADD_OBJECT = "AddObject",
    REMOVE_OBJECT = "RemoveObject",
    SET_POSITION = "SetPosition",
    SET_ROTATION = "SetRotation",
    SET_SCALE = "SetScale",
    SET_MATERIAL_COLOR = "SetMaterialColor",
    SET_MATERIAL_VALUE = "SetMaterialValue",
    SET_GEOMETRY = "SetGeometry",
    MUTLI_CMDS = "MultiCmds",
    ATTACH_BEHAVIOR = "AttachBehavior",
    DETACH_BEHAVIOR = "DetachBehavior",
    UPDATE_BEHAVIOR = "UpdateBehavior",
    GENERATE_3D_OBJECT = "Generate3dObject",
    ADD_3D_OBJECT = "Add3dObject",
    GENERATE_ERTH_MODEL = "GenerateErthModel",
    SET_MATERIAL_TEXTURE = "SetMaterialTexture",
    COMPLETE = "Complete",
    // AI contextual commands
    GET_PLAYER_DATA = "GetPlayerData",
    GET_SCENE_DATA = "GetSceneData",
    GET_SELECTED_OBJECT_DATA = "GetSelectedObjectData",
    GET_OBJECT_DATA = "GetObjectData",
    GET_LOOK_AT_POINT = "GetLookAtPoint",
    GET_SEARCH_RESULTS = "GetSearchResults",
    GET_BEHAVIORS_CONFIG = "GetBehaviorsConfig",
}
export enum AI_AGENT_MODE {
    EDITOR = "editor",
    SANDBOX_GENERATION = "sandbox_generation",
}
export type AiCommand = {
    type: COMMANDS;
    params: any;
    requiresUserConfirmation?: boolean;
};

export type AiCommandsResponse = {
    response: string;
    threadId?: string;
    commands: AiCommand;
};

export type AiAgentRequest = {
    userMessage: string;
    params: {
        sceneData?: string;
        playerData?: string;
        selectedObjectData?: string;
        objectData?: string;
        behaviorConfig?: string;
        playerWidth?: string;
        playerHeight?: string;
        docs?: string;
        starterCode?: string;
        lookAtPointData?: string;
        searchResults?: string;
    };
};

export type PendingCommandData = {
    aiCommand: AiCommand | null;
    command: unknown;
    aiAgentRequest: AiAgentRequest | null;
};